+++ src/account/operations/syncing/addresses/output_ids/mod.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod alias;
+mod alias_foundry;
 mod basic;
 mod nft;
